:heavy_plus_sign:Refactoring 의존성 주입.
